Improve rocketchat (#6)
* Improve rocketchat reporter

* Reduce code duplication

* Moved test to clientutil
